[client] don't send credentials when connecting to API proxy server
https://github.com/browserify/http-browserify/blob/1.7.0/lib/request.js#L18-L23

The reason for this update is:
  * when 'http' is replaced by 'http-browserify'
    - XMLHttpRequest.withCredentials defaults to TRUE
    - response header 'Access-Control-Allow-Origin: *' is not allowed
  * in the original API Proxy Server
    - calling: api({allowOrigin: true})
      adds this wildcard response header to all API endpoints,
      which only results in a client-side warning

The API proxy server in this fork already provides a solution.
The CLI option: -o true
results in the call: api({allowOrigin: true})
but api now processes a boolean value of TRUE differently.
Rathen than using a wildcard, this special value
causes the value of the response header: 'Access-Control-Allow-Origin'
to exactly match the origin of each individual request,